src/third_party/blink/web_tests/external/wpt/html/webappapis/system-state-and-capabilities/the-navigator-object/historical.https.window-expected.txt

This is a testharness.js-based test.
PASS registerContentHandler() is removed
PASS isProtocolHandlerRegistered() is removed
PASS isContentHandlerRegistered() is removed
PASS unregisterContentHandler() is removed
FAIL registerProtocolHandler has no third argument assert_false: expected false got true
Harness: the test ran to completion.
